+++ test/demo.cpp
@@ -15,9 +15,11 @@ int main(int argc, char** argv) {
         USER_DICT_PATH);
   vector<string> words;
   vector<cppjieba::Word> jiebawords;
+  string s;
   string result;
-  string s = "我是拖拉机学院手扶拖拉机专业的。不用多久，我就会升职加薪，当上CEO，走上人生巅峰。";
 
+  s = "他来到了网易杭研大厦";
+  cout << s << endl;
   cout << "[demo] Cut With HMM" << endl;
   jieba.Cut(s, words, true);
   cout << limonp::Join(words.begin(), words.end(), "/") << endl;
@@ -26,10 +28,14 @@ int main(int argc, char** argv) {
   jieba.Cut(s, words, false);
   cout << limonp::Join(words.begin(), words.end(), "/") << endl;
 
+  s = "我来到北京清华大学";
+  cout << s << endl;
   cout << "[demo] CutAll" << endl;
   jieba.CutAll(s, words);
   cout << limonp::Join(words.begin(), words.end(), "/") << endl;
 
+  s = "小明硕士毕业于中国科学院计算所，后在日本京都大学深造";
+  cout << s << endl;
   cout << "[demo] CutForSearch" << endl;
   jieba.CutForSearch(s, words);
   cout << limonp::Join(words.begin(), words.end(), "/") << endl;
@@ -42,12 +48,12 @@ int main(int argc, char** argv) {
   cout << limonp::Join(words.begin(), words.end(), "/") << endl;
 
   cout << "[demo] CutForSearch Word With Offset" << endl;
-  jieba.SetQuerySegmentThreshold(3);
-  jieba.CutForSearch("南京市长江大桥", jiebawords, true);
+  jieba.CutForSearch(s, jiebawords, true);
   cout << jiebawords << endl;
 
   cout << "[demo] Tagging" << endl;
   vector<pair<string, string> > tagres;
+  s = "我是拖拉机学院手扶拖拉机专业的。不用多久，我就会升职加薪，当上CEO，走上人生巅峰。";
   jieba.Tag(s, tagres);
   cout << s << endl;
   cout << tagres << endl;;